refactor: isolate daemon overlay and merge helpers
